requirements.txt
extra_streamlit_components==0.1.56
numpy==1.23.4
opencv_python==4.6.0.66
pandas==1.5.0
Pillow==9.3.0
PyJWT==2.6.0
python_bcrypt==0.3.2
PyYAML==6.0
seaborn==0.12.1
streamlit==1.13.0
streamlit_authenticator==0.2.1
tensorflow_gpu==2.8.0
